test: Remove unnecessary test
